pyproject.toml
[project]
name = "gaia-agent"
version = "0.1.0"
description = "GAIA agent using LangGraph"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4.1",
    "langchain-anthropic>=0.3.0",
    "python-dotenv>=1.0.0",
    "langchain-tavily>=0.2.17",
    "langfuse>=4.0.1",
    "pyyaml>=6.0.3",
    "e2b-code-interpreter>=2.6.0",
    "docling>=2.31.0",
]

[project.scripts]
print-mermaid = "scripts.print_mermaid_for_agent_graph:main"
evaluate-agent = "scripts.evaluate_agent:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["evaluate_agent_on_dataset"]

[tool.setuptools.packages.find]
where = ["src", "."]
include = ["agent_graph*", "evaluators*", "tools*", "scripts*"]

[tool.setuptools.package-dir]
"" = "src"
"scripts" = "scripts"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]

[tool.uv]
required-environments = [
    "sys_platform == 'darwin' and platform_machine == 'x86_64'"
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
